test/inspector/debugger/compile-function-expected.txt

compile function

Running test: testFailCompileFunction
hasBreakpoint: false

Running test: testSuccessCompileFunction
hasBreakpoint: true
